backend/requirements.txt
azure-monitor-opentelemetry==1.6.4
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.1
sqlalchemy==2.0.31
alembic==1.13.2
asyncpg==0.29.0
redis==5.0.7
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
